Automatically update wpa3 only setting in QR code
Replace ```wpa3_support``` and ```wpa3_transition``` image attributes with ```wpa_mode```. It is used to decide between ```WPA2```, ```WPA2/WPA3```, and ```WPA3``` based on the now defunct attributes' source keys in wlanconf. This new attribute will be used to determine if an additional flag is needed in the QR code for WPA3 only mode.