Fix the qr_text state attribute
The state attribute ```wpa_mode``` was being used to determine the correct generation string to convert into a QR code. For currently unknown reasons, this attribute was proving unreliable. It would default to WPA3 which caused an additional flag to be included in the generation string. The generation string is available as the state attribute ```qr_text```